+++ test-frontend/src/app/page.tsx
@@ -1,5 +1,4 @@
 import PanelClients from "@/components/panel-clients";
-import Image from "next/image";
 
 export default function Home() {
   return (
