+++ test-frontend/src/types/types.d.ts
@@ -1,0 +1,15 @@
+enum Roles {
+    disabled = 1,
+    active = 2,
+    inactive = 3,
+    waiting = 4,
+  }
+  
+  type Client = {
+    id: number;
+    displayName: string;
+    email: string;
+    document: string;
+    phone: string;
+    roleId: Roles;
+  };
